Responsiveness + Folders + searchBar
Hey, I added folders to the main page, made sure they don't affect the navbar or the footer, made this section responsive no matter how many folders are being added, fixed the footer not sticking to the bottom, fixed the div not aligning well, added the search bar, and I also put an option to add the folders dynamically. It's not yet cleaned up well but I plan to do so after working on the backend. The backend needs a lot of work because the login was hard coded, which is why the token was returning garbage and not redirecting smoothly. I will work on it and update you so  I can finalize the project and allow the admin to add/delete folders as well as add files to the folder to then filter as freely as they wish.